+++ CHANGELOG.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [0.1.0] - 2025-05-09 - Vlad
+### Added
+- Sending documents function
+
 ## [0.1.0] - 2025-05-08 - Vlad
 ### Added
 - SQLAlchemy repositories functionality